python/pyproject.toml
[project]
name = "hooteproto"
version = "0.1.0"
description = "Python client for Hootenanny"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyzmq>=26.0.0",
    "msgpack>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
hooteproto-gen = "hooteproto.gen:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hooteproto"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
